Rename the CI_X* defines Make the ci_service_xdata.xopts an uint64_t integer add the stdint.h header to c-icap.h
git-svn-id: https://svn.code.sf.net/p/c-icap/code/c-icap-server/trunk/c-icap@215 950cb5d8-81e4-40ab-8cb8-c719d5138fe1

+++ service_lib.c
@@ -86,14 +86,14 @@ void ci_service_enable_204(ci_service_xdata_t * srv_xdata)
 }
 
 
-void ci_service_set_xopts(ci_service_xdata_t * srv_xdata, int xopts)
+void ci_service_set_xopts(ci_service_xdata_t * srv_xdata, uint64_t xopts)
 {
      ci_thread_rwlock_wrlock(&srv_xdata->lock);
      srv_xdata->xopts = xopts;
      ci_thread_rwlock_unlock(&srv_xdata->lock);
 }
 
-void ci_service_add_xopts(ci_service_xdata_t * srv_xdata, int xopts)
+void ci_service_add_xopts(ci_service_xdata_t * srv_xdata, uint64_t xopts)
 {
      ci_thread_rwlock_wrlock(&srv_xdata->lock);
      srv_xdata->xopts |= xopts;
